Implemented mqtt for app and WIO terminal #22
Created a barebone app with mqtt functionality and a corresponding
subtopic for each sensor reading to be able to update values on the
screen.

For the WIO terminal there is also the implementation of mqtt.
However, it needs to be refactored and restructured to work with
the rest of the code.