Implements #2
Revisited almost the entirety of the WIO terminals MqttClient to make it
work more nicely with other parts of the system. This translates to:
less system wide delays.

Known issues:
Whenever the WIO attempts to reconnect to the mqtt broker it will be a
system wide pause until it reconnects. This can unfortunately not be
fixed, as it is how the mqtt library we use implementation.

Other changes:
1. Revamped buzzer functionality to better work as an alert.
2. Added Util class to manage time management. This to not have the
current time of the system differentiate due to computation time.
3. Renamed some variables & improved some comments.