Implements #62
This commit implements the following changes:
- The app now uses a universal BrokerConnection object,
which is created in the App class and can be accessed from any activity
instead of instantiating a new BrokerConnection object in each activity.
This solved the issue of the tests requiring access to the connection
state without having it static as that would break the mqqt.

- Deleted a lot of redundant codes and files that were not used anymore.
These were mostly tied to the old design of the app and were therefore
not needed anymore.

- General code cleanup and refactoring.

Things to still improve:
- The app will display the reconnect activity multiple times if the
user changes screens while the app is not connected. This is because
whenever a new activity is created, it checks if the app is connected
and if not, it will display the reconnect activity after one second.
During this one second, the user can change screens, which will create
a new activity and therefore display the reconnect activity again.

- Could also store current values for people, loudness, temperature
and humidity in the app. Since right now if the WIO is not connected,
the app will not display the last know values if you change screens.